Extract the shared platform axis into platform.hpp
The theme foundation and the keymap core each declared their own
PlatformId enum and a platformIdName helper, two copies of one axis. Move
both into a small Qt-free platform.hpp that each core includes, so they
name the same platforms and the running-platform detection resolves
against a single definition.
